Make it possible for stick movement to affect just one axis (rather than both X and Y at once)
That should allow the controller's sticks to achieve diagonal states by using commands such as:

- move left & move up
- custom([Lup; Lleft]; 0.5)
- onward, move left/right

(Note: onward is a command that CommunityController has supported in a few games.
Its actual meaning is to hold the left stick upward until the next time
it is reset to neutral - e.g. by the "stop" or "still" command.)